添加了HoloScore装饰品 添加了3D渲染相关

+++ RandomBuff/Core/Game/Settings/Conditions/CardCondition.cs
@@ -15,7 +15,7 @@ namespace RandomBuff.Core.Game.Settings.Conditions
     {
         public override ConditionID ID => ConditionID.Card;
 
-        public override float Exp => needCard * 50;
+        public override int Exp => needCard * 50;
 
         public override void GachaEnd(List<BuffID> picked, List<BuffID> allCards)
         {
